Rename index.js to childchain.js and rootchain.js
