+++ packages/omg-js-childchain/package.json
@@ -7,7 +7,7 @@
     "OmiseGo",
     "Plasma"
   ],
-  "main": "src/index.js",
+  "main": "src/childchain.js",
   "author": "OmiseGo",
   "contributors": [
     "Pong Cheecharern <pong@omise.co>",
