+++ packages/omg-js-childchain/src/childchain.js
