+++ packages/omg-js-childchain/test/getBalanceTest.js
@@ -15,7 +15,7 @@ limitations under the License. */
 
 const chai = require('chai')
 const chaiAsPromised = require('chai-as-promised')
-const ChildChain = require('../src')
+const ChildChain = require('../src/childchain')
 const nock = require('nock')
 
 chai.use(chaiAsPromised)
