+++ packages/omg-js-rootchain/src/rootchain.js
